remove legacy support for XRD and host-meta
This removes support for host-meta as well as the XRD version of
webfinger documents.  Other plugins can still support alternate
representations of the webfinger resource using the 'webfinger_render'
action.  This action has been simplified to take only the webfinger data
array, since there shouldn't be anything user-specific in the rendering,
and the HTTP Accept header can be retrieved from $_SERVER if needed.

The default behavior will be to render the webfinger resource as a JRD,
regardless of the HTTP Accept header.  This is consistent with the
WebFinger specs instruction to ignore unsupported representations.

This commit also switches to use WordPress's status_header() function.

fixes #1